refactor: Move test counter server into .py file
