+++ cmd/unikraft/integration/instance_test.go
@@ -6,6 +6,7 @@
 package integration
 
 import (
+	_ "embed"
 	"strings"
 	"testing"
 	"time"
@@ -16,6 +17,9 @@ import (
 
 	integ "unikraft.com/cli/internal/integration"
 )
+
+//go:embed testdata/counter_server.py
+var counterServerPy []byte
 
 func TestInstances(t *testing.T) {
 	t.Run("create", func(t *testing.T) {
@@ -1200,46 +1204,7 @@ func applyCounterContext(dir string) error {
 FROM python:3.12-slim
 COPY server.py /app/server.py
 `), 0o644),
-		fstest.CreateFile("server.py", []byte(`
-import json
-from http.server import HTTPServer, BaseHTTPRequestHandler
-
-counter = 0
-
-class Handler(BaseHTTPRequestHandler):
-    def do_GET(self):
-        if self.path == "/count":
-            self.send_response(200)
-            self.send_header("Content-Type", "application/json")
-            self.end_headers()
-            self.wfile.write(json.dumps({"count": counter}).encode())
-        else:
-            self.send_response(404)
-            self.end_headers()
-
-    def do_POST(self):
-        global counter
-        if self.path == "/increment":
-            length = int(self.headers.get("Content-Length", 0))
-            body = json.loads(self.rfile.read(length)) if length else {}
-            delta = body.get("delta", 1)
-            if delta == "reset":
-                counter = 0
-            else:
-                counter += int(delta)
-            self.send_response(200)
-            self.send_header("Content-Type", "application/json")
-            self.end_headers()
-            self.wfile.write(json.dumps({"count": counter}).encode())
-        else:
-            self.send_response(404)
-            self.end_headers()
-
-    def log_message(self, format, *args):
-        pass  # silence request logging
-
-HTTPServer(("", 8080), Handler).serve_forever()
-`), 0o644),
+		fstest.CreateFile("server.py", counterServerPy, 0o644),
 		fstest.CreateFile("Kraftfile", []byte(`
 spec: v0.7
 name: counter-e2e
